Move patch into own file
