+++ lib/broken_record/job_result.rb
@@ -22,11 +22,11 @@ module BrokenRecord
       @original_errors << [id, errors] if errors
       if exception
         exception_hash = {
-            context: exception.backtrace.grep(Regexp.new(Rails.root.to_s))[0].gsub("#{Rails.root}/", ''),
-            exception_class: exception.is_a?(Class) ? exception : exception.class,
-            message: exception.message,
-            source: exception.backtrace
-          }
+          context: exception.backtrace.grep(Regexp.new(Rails.root.to_s))[0].gsub("#{Rails.root}/", ''),
+          exception_class: exception.is_a?(Class) ? exception : exception.class,
+          message: exception.message,
+          source: exception.backtrace
+        }
         @exceptions << [id, exception_hash]
       end
     end
